// admin_live_mall/src/pages/Data/DealData/TransactionOverview/data.d.ts
export interface transactionDataType {
  payedNum: number;
  payedUserNum: string;
  pv: string;
  refundAmount: string;
  regMemberNum: string;
  totalAmount: string;
  turnover: string;
  unitAmount: string;
  uv: string;
}
export interface tranComputerDataType {
  payedNum: string;
  payedUserNum: string;
  pv: string;
  refundAmount: string;
  regMemberNum: string;
  totalAmount: string;
  turnover: string;
  unitAmount: string;
  uv: string;
}